Trait future_queue::traits::GroupedWeightedFuture

pub trait GroupedWeightedFuture: Sealed {
    type Future: Future;
    type Q;

    // Required methods
    fn weight(&self) -> usize;
    fn into_components(self) -> (usize, Option<Self::Q>, Self::Future);
}

A trait for types which can be converted into a Future, an optional group, and a weight.

Provided in case it’s necessary. This trait is only implemented for (usize, Option<Q>, impl Future).

Required Associated Types

type Future: Future

The associated Future type.

type Q

The associated key lookup type.

Required Methods

fn weight(&self) -> usize

Returns the weight.

fn into_components(self) -> (usize, Option<Self::Q>, Self::Future)

Turns self into its components.