future_queue::traits

Trait GroupedWeightedFuture

pub trait GroupedWeightedFuture: Sealed {
    type F: FnOnce(FutureQueueContext) -> Self::Future;
    type Future: Future;
    type Q;

    // Required methods
    fn weight(&self) -> usize;
    fn into_components(self) -> (usize, Option<Self::Q>, Self::F);
}

A trait for types which can be converted into functions that return a Future, an optional group, and a weight.

Provided in case it’s necessary. This trait is only implemented for (usize, Option<Q>, impl Future).

Required Associated Types

type F: FnOnce(FutureQueueContext) -> Self::Future

The function to obtain the future from.

type Future: Future

The associated Future type.

type Q

The associated key lookup type.

Required Methods

fn weight(&self) -> usize

Returns the weight.

fn into_components(self) -> (usize, Option<Self::Q>, Self::F)

Turns self into its components.